returned app-development to its former glory and moved the shared modeling logic into a common directory
