+++ src/studio/src/designer/frontend/dashboard/tsconfig.json
@@ -21,6 +21,9 @@
     "paths": {
       "app-shared/*": [
         "../shared/*"
+      ],
+      "@altinn/schema-editor/*": [
+        "../packages/schema-editor/src/*"
       ]
     }
   },
