+++ src/studio/src/designer/frontend/dashboard/webpack.config.development.js
@@ -20,7 +20,8 @@ module.exports = {
   resolve: {
     extensions: [".ts", ".tsx", ".js", ".jsx", ".css", ".scss"],
     alias: {
-      "app-shared": path.resolve(__dirname, "../shared/")
+      "app-shared": path.resolve(__dirname, "../shared/"),
+      "@altinn/schema-editor": path.resolve(__dirname, "../packages/schema-editor/src/"),
     }
   },
   performance: {
@@ -28,56 +29,56 @@ module.exports = {
   },
   module: {
     rules: [{
-        test: /\.jsx?/,
-        exclude: /node_modules/,
-        use: {
-          loader: "babel-loader",
-        }
-      },
-      {
-        test: /\.html$/,
-        use: [{
-          loader: "html-loader",
-          options: {
-            minimize: true
-          }
-        }]
-      },
-      {
-        test: /\.scss$/,
-        use: [
-          "style-loader",
-          "css-loader"
-        ]
-      },
-      {
-        test: /\.svg$/,
-        use: {
-          loader: "svg-inline-loader",
-        }
-      },
-      {
-        test: /\.css$/,
-        use: [{
-            loader: MiniCssExtractPlugin.loader,
-          },
-          {
-            loader: "css-loader",
-            options: {
-              url: false
-            }
-          },
-        ]
-      },
-      {
-        test: /\.tsx?/,
-        loader: "ts-loader",
-      },
-      {
-        enforce: "pre",
-        test: /\.js$/,
-        loader: "source-map-loader",
+      test: /\.jsx?/,
+      exclude: /node_modules/,
+      use: {
+        loader: "babel-loader",
       }
+    },
+    {
+      test: /\.html$/,
+      use: [{
+        loader: "html-loader",
+        options: {
+          minimize: true
+        }
+      }]
+    },
+    {
+      test: /\.scss$/,
+      use: [
+        "style-loader",
+        "css-loader"
+      ]
+    },
+    {
+      test: /\.svg$/,
+      use: {
+        loader: "svg-inline-loader",
+      }
+    },
+    {
+      test: /\.css$/,
+      use: [{
+        loader: MiniCssExtractPlugin.loader,
+      },
+      {
+        loader: "css-loader",
+        options: {
+          url: false
+        }
+      },
+      ]
+    },
+    {
+      test: /\.tsx?/,
+      loader: "ts-loader",
+    },
+    {
+      enforce: "pre",
+      test: /\.js$/,
+      loader: "source-map-loader",
+    }
     ],
   },
   plugins: [
