+++ src/studio/src/designer/frontend/shared/features/index.ts
@@ -1,0 +1,1 @@
+export { default as DataModelling } from "./dataModelling";
